[core] Split-off POSIX sync version into a separate file (#1455)
