+++ srtcore/sync.h
@@ -729,7 +729,7 @@ public:
     /// To be used only in StartThread function.
     /// Creates a new stread and assigns to this.
     /// @throw CThreadException
-    inline void create_thread(void *(*start_routine) (void *), void *arg);
+    void create_thread(void *(*start_routine) (void *), void *arg);
 #endif
 
 public: // Observers
